Move NetworkInterface class to dedicated module
